app/raspberry/requirements.txt
paho-mqtt>=1.6.1
websockets>=12.0

